Add the most basic of a Graph structure
My current thoughts are to use the graphviz algorithm to draw general
graphs, and we'll use that to model all of our data structure
visualizations.

The representation I've gone with here is an adjacency list, but we
should try and keep the graph representation as generic as possible. We
might need to be able to support edge weights later, but not sure how
exactly we should do that yet.